relocated theme components
