+++ packages/base-components-mui-rne/src/components/ThemeHeadline1.tsx
@@ -1,0 +1,14 @@
+import React, { FunctionComponent } from 'react';
+import ThemeTextStyles from "./ThemeTextStyles";
+import ThemeText from "./ThemeText";
+import ThemeTextProps from "../types/ThemeText";
+
+export const ThemeHeadline1: FunctionComponent<ThemeTextProps> = ({children, textStyle})=>{
+    return (
+        <ThemeText textStyle={[ThemeTextStyles.h1, textStyle]}>
+        {children}
+        </ThemeText>
+    )
+}
+
+export default ThemeHeadline1
