+++ packages/base-components-mui-rne/src/components/ThemeHeadline2.tsx
@@ -1,0 +1,14 @@
+import React, { FunctionComponent } from 'react';
+import ThemeTextStyles from "./ThemeTextStyles";
+import ThemeText from "./ThemeText";
+import ThemeTextProps from "../types/ThemeText";
+
+export const ThemeHeadline2: FunctionComponent<ThemeTextProps> = ({children, textStyle})=>{
+    return (
+        <ThemeText textStyle={[ThemeTextStyles.h2, textStyle]}>
+        {children}
+        </ThemeText>
+    )
+}
+
+export default ThemeHeadline2
