+++ packages/base-components-mui-rne/src/components/ThemeText.tsx
@@ -1,8 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import { Text } from 'react-native-elements';
-import ThemeTextProps from '../types/ThemeText'
+import ThemeTextProps from "../types/ThemeText";
 
-export const ThemeText: FunctionComponent<ThemeTextProps> = ({textStyle, children})=>{
+export const ThemeText: FunctionComponent<ThemeTextProps> = ({children, textStyle})=>{
     return(
         <Text style = {textStyle}>
             {children}
