+++ packages/base-components-mui-rne/src/components/ThemeTextButton.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react';
 import { Button } from 'react-native-elements';
-import ThemeTextButtonProps from '../types/ThemeTextButton'
+import ThemeTextButtonProps from "../types/ThemeTextButton";
 
 export const ThemeTextButton: FunctionComponent<ThemeTextButtonProps> = ({ disabled, onClickOrPress, title }) => {
   return (
