+++ packages/base-components-mui-rne/src/components/ThemeWarningMessage.tsx
@@ -1,0 +1,13 @@
+import React, { FunctionComponent } from 'react';
+import ThemeText from "./ThemeText";
+import ThemeTextProps from "../types/ThemeText";
+import ThemeTextStyles from "./ThemeTextStyles";
+
+
+export const ThemeWarningMessage: FunctionComponent<ThemeTextProps> = ({children, textStyle}) => { 
+    return <ThemeText textStyle={[ThemeTextStyles.warningMessage, textStyle]}>
+        {children}
+    </ThemeText>
+}
+
+export default ThemeWarningMessage
