+++ packages/base-components-mui-rne/src/components/index.ts
@@ -1,8 +1,8 @@
-export {ThemeHeadline1} from '../../../common/src/shared/components/ThemeHeadline1'
-export {ThemeHeadline2} from '../../../common/src/shared/components/ThemeHeadline2'
-export {ThemeParagraph} from '../../../common/src/shared/components/ThemeParagraph'
-export {ThemeTextButton} from '../../../common/src/shared/components/ThemeTextButton'
-export {ThemeTextField} from '../../../common/src/shared/components/ThemeTextField'
-export {ThemeTextLink} from '../../../common/src/shared/components/ThemeTextLink'
-export {ThemeErrorMessage} from '../../../common/src/shared/components/ThemeErrorMessage'
-export {ThemeWarningMessage} from '../../../common/src/shared/components/ThemeWarningMessage'
+export {ThemeHeadline1} from "./ThemeHeadline1"
+export {ThemeHeadline2} from "./ThemeHeadline2"
+export {ThemeParagraph} from "./ThemeParagraph"
+export {ThemeTextButton} from "./ThemeTextButton"
+export {ThemeTextField} from "./ThemeTextField"
+export {ThemeTextLink} from "./ThemeTextLink"
+export {ThemeErrorMessage} from "./ThemeErrorMessage"
+export {ThemeWarningMessage} from "./ThemeWarningMessage"
