+++ packages/base-components-mui-rne/src/types/ThemeText.ts
@@ -1,0 +1,7 @@
+import {StyleProp, TextStyle} from 'react-native';
+
+type ThemeTextProps = {
+    textStyle?: StyleProp<TextStyle>
+}
+
+export default ThemeTextProps
