+++ packages/base-components-mui-rne/src/types/ThemeTextButton.ts
@@ -1,8 +1,7 @@
-
 type ThemeTextButtonProps = {
   disabled?: boolean,
   onClickOrPress: any,
   title: string
 }
 
-export default ThemeTextButtonProps;
+export default ThemeTextButtonProps
