+++ packages/base-components-mui-rne/src/types/ThemeTextLink.ts
@@ -1,0 +1,4 @@
+type ThemeTextLinkProps = {
+    toPath: string
+}
+export default ThemeTextLinkProps
