--- packages/common/src/shared/components/ThemeHeadlineStyles.ts
@@ -1,14 +1,0 @@
-import {StyleSheet} from 'react-native'
-
-const ThemeHeadlineStyles = StyleSheet.create({
-    h1 :{
-        fontSize:50,
-        fontWeight:"400",
-    },
-    h2 :{
-        fontSize:25,
-        fontWeight:"200",
-    }
-})
-
-export default ThemeHeadlineStyles
