--- packages/common/src/shared/types/ThemeMessage.ts
@@ -1,6 +1,0 @@
-type ThemeMessageProps = {
-    props?:any
-    // could be used in future
-}
-
-export default ThemeMessageProps
