--- packages/common/src/shared/types/ThemeText.ts
@@ -1,7 +1,0 @@
-import {StyleProp, TextStyle} from 'react-native'
-
-type ThemeTextProps = {
-    textStyle?: StyleProp<TextStyle>
-}
-
-export default ThemeTextProps
