--- packages/common/src/shared/types/ThemeTextLink.ts
@@ -1,4 +1,0 @@
-type ThemeTextLinkProps = {
-    to: string
-}
-export default ThemeTextLinkProps
